Create README.md
Title of the Project: Form Validation

Description: HTML form validation is a process of examining the HTML form page's contents to avoid errored-out data being sent to the server. This process is a significant step in developing HTML-based web applications, as it can easily improve the quality of the web page or the web application.

Benefits of using JsonPowerDB:
It is used for data transmission between a server and web applications
JSON is much easier to parse. But this is only relevant if one is writing a parser which is not a common activity at this point.
JSON syntax is very easy to use

Project status: Successfully done with the database connectivity of the Student database (Student-Rel) and created a validation form.

Sources: Login2Xplore provided video